q!ltc to /ltc
